陽性者の属性: Rename components/DataTable.vue to components/index/CardsReference/ConfirmedCasesAttributes/DataTable.vue
