+++ components/index/CardsReference/ConfirmedCasesAttributes/DataTable.vue
